Killed all compiler warnings. No more need for a dummy stpcpy() function as this is now implemented in clib.
git-svn-id: https://svn.aros.org/svn/aros/trunk@11385 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/aminet/util/dir/filemaster/fmgui.c
@@ -658,7 +658,7 @@ void buttonbari(struct GUIBase *gb,void **buttons)
 {
 startbuttonbar(gb);
 while(*buttons) {
-	reqbutton(gb,buttons[0],(WORD)buttons[1],guiUC);
+	reqbutton(gb,buttons[0],(LONG)buttons[1],guiUC);
 	buttons+=2;
 }
 }
